+++ packages/superdoc/src/components/CommentsLayer/CommentDialog.vue
@@ -6,7 +6,6 @@ import { useCommentsStore } from '@superdoc/stores/comments-store';
 import { useSuperdocStore } from '@superdoc/stores/superdoc-store';
 import { SuperInput } from '@harbour-enterprises/super-editor';
 import { superdocIcons } from '@superdoc/icons.js';
-import { isAllowed, PERMISSIONS } from '@superdoc/core/collaboration/permissions.js';
 import useSelection from '@superdoc/helpers/use-selection';
 import useComment from '@superdoc/components/CommentsLayer/use-comment';
 import Avatar from '@superdoc/components/general/Avatar.vue';